Hotstrings Export format selection now in File Save Dialog
